API CHANGE, replace auto-height with grid-height
New attribute to allow define the height of actual grid inside
px-data-grid. Currently supported values are undefined and
'default' both just using the default heigth of vaadin-grid. Then
'auto' that is as old auto-height. And then any non-relative
CSS height value (eg. '300px'). This does not yet support pairing
sizing with define height of px-data-grid. That is new requirement
and needs to be done later.

Fixes #333