Cancel tap and click events
Internally, px-dropdown is listening for an event called 'tap' on
its trigger. That's for legacy reasons. 'tap' is a Polymer 1.x
concept that was meant to normalize click events in touch
and mouse environments. To maintain compatability going forward
with px-dropdown it will continue to use the 'tap' event instead
of the 'click' event.

Because of this, the data grid is cancelling the click event but
allowing the tap event to bubble, which leads to the grid
opening the dropdown through property binding and the dropdown
toggling itself closed immediately.

This fix ensures that both the click and tap events are cancelled
and allows the dropdown to open correctly.